+++ src/constants/Config.ts
@@ -2,6 +2,9 @@ export const Config = {
   root: "demoTime",
   highlight: {
     borderColor: "highlightBorderColor",
+    background: "highlightBackground",
+    blur: "highlightBlur",
+    opacity: "highlightOpacity",
     zoom: "highlightZoomEnabled",
   },
   clock: {